image_processing/requirements.processor.txt
# ===== requirements.processor.txt =====
# Image Processor Requirements - Ubuntu/Debian compatible
# Image processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# Data processing
PyYAML>=6.0
tqdm>=4.65.0

# Database connectivity
pyodbc>=4.0.39
jpype1>=1.4.0
jaydebeapi>=1.2.3

# Image metadata
exifread>=3.0.0
piexif>=1.1.3

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Utilities
pydantic>=2.4.0
structlog>=23.1.0
python-dateutil>=2.8.0

# HTTP client
httpx>=0.24.0
requests>=2.31.0